+++ render_to_markdown.py
@@ -2,7 +2,6 @@ import io
 import re
 from datetime import datetime
 from typing import Generator
-from copy import deepcopy
 
 from tqdm import tqdm
 
@@ -57,14 +56,11 @@ def _clean_article_html(
 def _is_tm_in_force(tm: CodeJSON, timestamp: int) -> bool:
     if tm.get("nature", None) == "CODE":
         return True  # root
-    start_timestamp = int(
-        datetime.fromisoformat(f"{tm['dateDebut']} 00:00").timestamp() * 1000
-    )
-    end_timestamp = int(
-        datetime.fromisoformat(f"{tm['dateFin']} 23:59").timestamp() * 1000
-    )
 
-    return start_timestamp <= timestamp <= end_timestamp
+    start_timestamp = datetime.fromisoformat(f"{tm['dateDebut']} 00:00").timestamp()
+    end_timestamp = datetime.fromisoformat(f"{tm['dateFin']} 23:59").timestamp()
+
+    return start_timestamp <= timestamp / 1000 <= end_timestamp
 
 
 def _tm_to_markdown(
@@ -98,62 +94,6 @@ def _tm_to_markdown(
     # TODO
 
 
-def _get_tm_by_path(tm: CodeJSON, path: [str]) -> CodeJSON:
-    if len(path) == 0:
-        return tm
-
-    for section in tm["sections"]:
-        if section["cid"] == path[0]:
-            return _get_tm_by_path(section, path[1:])
-
-    raise KeyError(f"Section {path} not found in tm")
-
-
-def _format_path(titresTm):
-    return "/".join([t["cid"] for t in titresTm])
-
-
-def _test_paths(tm: CodeJSON, paths: [str], article_cid: str) -> ([str], [str]):
-    valid = []
-    missing = []
-
-    for path in paths:
-        try:
-            section = _get_tm_by_path(tm, path.split("/"))
-            found = [a for a in section["articles"] if a["cid"] == article_cid]
-            assert len(found) == 1
-            valid.append(path)
-        except AssertionError:
-            missing.append(path)
-
-    return valid, missing
-
-
-def _patch_tm_multiple_paths(tm: CodeJSON, articles: [ArticleJSON]) -> CodeJSON:
-    patched_tm = deepcopy(tm)
-
-    for article in articles:
-        v_0 = article["listArticle"][0]
-        if v_0["textTitles"][0]["cid"] == tm["cid"]:
-            paths = {
-                _format_path(v["context"]["titresTM"]) for v in article["listArticle"]
-            }
-            if len(paths) > 1:
-                valid, missing = _test_paths(tm, paths, v_0["cid"])
-                article_ref = next(
-                    a
-                    for a in _get_tm_by_path(tm, valid[0].split("/"))["articles"]
-                    if a["cid"] == v_0["cid"]
-                )
-                for m in missing:
-                    _get_tm_by_path(patched_tm, m.split("/"))["articles"] = sorted(
-                        _get_tm_by_path(patched_tm, m.split("/"))["articles"]
-                        + [article_ref],
-                        key=lambda x: x["num"],
-                    )
-    return patched_tm
-
-
 def generate_markdown(
     code_tms: list[CodeJSON], articles: list[ArticleJSON], sorted_commits: list[Commit]
 ) -> Generator[StateAtCommit, None, None]:
@@ -183,12 +123,10 @@ def generate_markdown(
         full_code_texts = []
         for tm in code_tms:
             f = io.StringIO()
-
             print("=" * 6 + f" {i} " + "=" * 6)
             print(cleaned_commits[i])
             print("\n\n")
-            patched_tm = _patch_tm_multiple_paths(tm, articles)
-            _tm_to_markdown(patched_tm, cleaned_commits[: (i + 1)], file=f)
+            _tm_to_markdown(tm, cleaned_commits[: (i + 1)], file=f)
             full_code_texts.append((tm["title"], f.getvalue()))
 
         yield StateAtCommit(
